+++ src/settings_internal.js
@@ -137,6 +137,9 @@ var PTHREAD_WORKER_FILE = '';
 // name of the file containing the Wasm Worker *.ww.js, if relevant
 var WASM_WORKER_FILE = '';
 
+// name of the file containing the Audio Worklet *.aw.js, if relevant
+var AUDIO_WORKLET_FILE = '';
+
 // Base URL the source mapfile, if relevant
 var SOURCE_MAP_BASE = '';
 
@@ -165,6 +168,9 @@ var MINIFY_ASMJS_EXPORT_NAMES = true;
 
 // Internal: represents a browser version that is not supported at all.
 var TARGET_NOT_SUPPORTED = 0x7FFFFFFF;
+
+// Used to track whether target environment supports the 'globalThis' attribute.
+var SUPPORTS_GLOBALTHIS = false;
 
 // Wasm backend symbols that are considered system symbols and don't
 // have the normal C symbol name mangled applied (== prefix with an underscore)